+++ repository/src/main/java/org/grnet/status/repositories/DowntimeRepository.java
@@ -7,6 +7,8 @@ import org.grnet.status.entities.Page;
 import org.grnet.status.entities.PageQuery;
 import org.grnet.status.entities.PageQueryImpl;
 import org.grnet.status.entities.*;
+import org.grnet.status.entities.*;
+
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.StringJoiner;